Use codec derive serialization for data types
