+++ fuzz-util/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "fuzz-util"
+version = "0.0.0"
+authors = ["Automatically generated"]
+publish = false
+edition = "2018"
+
+[dependencies]
+elrond-wasm = "0.14.2"
+
+[lib]
