+++ node-storage/fuzz/Cargo.toml
@@ -12,13 +12,21 @@ cargo-fuzz = true
 [dependencies]
 libfuzzer-sys = "0.3"
 elrond-wasm = "0.14.2"
+elrond-wasm-derive = "0.14.2"
 
 [dependencies.node-storage]
 path = ".."
 
+[dependencies.fuzz-util]
+path = "../../fuzz-util"
+
 # Prevent this from interfering with workspaces
 [workspace]
 members = ["."]
+
+[lib]
+name = "old_serialization"
+path = "old_serialization/lib.rs"
 
 [[bin]]
 name = "fuzz_bls_key"
