+++ user-fund-storage/fuzz/old_serialization/mod.rs
@@ -1,0 +1,7 @@
+pub mod fund_item;
+pub mod fund_list_info;
+pub mod fund_type;
+
+pub use fund_item::*;
+pub use fund_list_info::*;
+pub use fund_type::*;
